requirements.txt
flask
flask-cors
langchain-openai
google-cloud-firestore
langchain-google-firestore
python-dotenv
